Use "commander" for  cli argv handling (#1195)
* Use "commander" for  cli argv handling

* Handle different scripts version forms and exits without a name given

* Revert comment about min supported node version

* Check sooner for the minimal node version

* Add travis test for node <4

* Parse stderr in node versions <4

+++ tasks/e2e.sh
@@ -14,6 +14,11 @@
 # Start in tasks/ even if run from root directory
 cd "$(dirname "$0")"
 
+# CLI and app temporary locations
+# http://unix.stackexchange.com/a/84980
+temp_cli_path=`mktemp -d 2>/dev/null || mktemp -d -t 'temp_cli_path'`
+temp_app_path=`mktemp -d 2>/dev/null || mktemp -d -t 'temp_app_path'`
+
 function cleanup {
   echo 'Cleaning up.'
   cd $root_path
@@ -53,14 +58,22 @@ set -x
 cd ..
 root_path=$PWD
 
+npm install
+
+# If the node version is < 4, the script should just give an error.
+if [ `node --version | sed -e 's/^v//' -e 's/\..\+//g'` -lt 4 ]
+then
+  cd $temp_app_path
+  err_output=`node "$root_path"/packages/create-react-app/index.js test-node-version 2>&1 > /dev/null || echo ''`
+  [[ $err_output =~ You\ are\ currently\ running\ Node\ v.+\ but\ create-react-app\ requires\ \>=4\. ]] && exit 0 || exit 1
+fi
+
 if [ "$USE_YARN" = "yes" ]
 then
   # Install Yarn so that the test can use it to install packages.
   npm install -g yarn@0.17.10 # TODO: remove version when https://github.com/yarnpkg/yarn/issues/2142 is fixed.
   yarn cache clean
 fi
-
-npm install
 
 # Lint own code
 ./node_modules/.bin/eslint --ignore-path .gitignore ./
@@ -117,13 +130,10 @@ mv package.json.orig package.json
 # ******************************************************************************
 
 # Install the CLI in a temporary location
-# http://unix.stackexchange.com/a/84980
-temp_cli_path=`mktemp -d 2>/dev/null || mktemp -d -t 'temp_cli_path'`
 cd $temp_cli_path
 npm install $cli_path
 
 # Install the app in a temporary location
-temp_app_path=`mktemp -d 2>/dev/null || mktemp -d -t 'temp_app_path'`
 cd $temp_app_path
 create_react_app --scripts-version=$scripts_path test-app
 
@@ -184,7 +194,6 @@ npm test -- --watch=no
 
 # Test the server
 npm start -- --smoke-test
-
 
 # ******************************************************************************
 # Test --scripts-version with a version number
